requirements.txt
mimesis==10.1.0
pydantic==1.10.8
pytest==7.3.1
requests==2.31.0
allure-pytest==2.13.2
jsonschema